+++ front_end/components/GlobalMessage.tsx
