+++ front_end/components/Header.tsx
@@ -5,8 +5,8 @@ import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { Link } from 'react-router-dom';
-import { userSignout } from './services/UserService';
-import { editUser } from './modules/UserModule';
+import { userSignout } from '../services/UserService';
+import { editUser } from '../modules/UserModule';
 import ToggleSideBar from './ToggleSideBar';
 
 const useStyles = makeStyles((theme: Theme) =>
